+++ tests/feature_selection/test_feature_selectors.py
@@ -7,7 +7,7 @@ from unittest import TestCase
 import pandas as pd
 import tsfresh.defaults
 import tsfresh.feature_selection.significance_tests
-import tsfresh.feature_selection.feature_selector
+import tsfresh.feature_selection.relevance
 
 # the unit tests in this module make sure that obvious irrelevant features are considered irrelevant by the feature
 # selection algorithms
@@ -35,8 +35,8 @@ class FeatureSelection(TestCase):
         p_value = tsfresh.feature_selection.significance_tests.target_binary_feature_binary_test(x, y)
         self.assertGreater(p_value, self.minimal_p_value_for_unsignificant_features)
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, True,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, True,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertGreater(result_series.p_value, self.minimal_p_value_for_unsignificant_features)
         self.assertEqual(result_series.type, "binary")
@@ -51,8 +51,8 @@ class FeatureSelection(TestCase):
 
         self.assertGreater(p_value, self.minimal_p_value_for_unsignificant_features)
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, True,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, True,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertGreater(result_series.p_value, self.minimal_p_value_for_unsignificant_features)
         self.assertEqual(result_series.type, "real")
@@ -67,8 +67,8 @@ class FeatureSelection(TestCase):
 
         self.assertGreater(p_value, self.minimal_p_value_for_unsignificant_features)
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, False,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, False,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertGreater(result_series.p_value, self.minimal_p_value_for_unsignificant_features)
         self.assertEqual(result_series.type, "binary")
@@ -83,8 +83,8 @@ class FeatureSelection(TestCase):
 
         self.assertGreater(p_value, self.minimal_p_value_for_unsignificant_features)
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, False,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, False,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertGreater(result_series.p_value, self.minimal_p_value_for_unsignificant_features)
         self.assertEqual(result_series.type, "real")
@@ -101,8 +101,8 @@ class FeatureSelection(TestCase):
         p_value = tsfresh.feature_selection.significance_tests.target_binary_feature_binary_test(x, y)
         self.assertLess(p_value, self.maximal_p_value_for_significant_features)
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, True,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, True,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertLess(result_series.p_value, self.maximal_p_value_for_significant_features)
         self.assertEqual(result_series.type, "binary")
@@ -122,8 +122,8 @@ class FeatureSelection(TestCase):
                                                                                                tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertLess(p_value, self.maximal_p_value_for_significant_features)
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, True,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, True,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertLess(result_series.p_value, self.maximal_p_value_for_significant_features)
         self.assertEqual(result_series.type, "real")
@@ -143,8 +143,8 @@ class FeatureSelection(TestCase):
         p_value = tsfresh.feature_selection.significance_tests.target_binary_feature_real_test(x, y, test="smir")
         self.assertLess(p_value, self.maximal_p_value_for_significant_features)
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, True,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, True,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertLess(result_series.p_value, self.maximal_p_value_for_significant_features)
         self.assertEqual(result_series.type, "real")
@@ -158,8 +158,8 @@ class FeatureSelection(TestCase):
         p_value = tsfresh.feature_selection.significance_tests.target_real_feature_binary_test(x, y)
         self.assertLess(p_value, self.maximal_p_value_for_significant_features)
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, False,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, False,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertLess(result_series.p_value, self.maximal_p_value_for_significant_features)
         self.assertEqual(result_series.type, "binary")
@@ -174,8 +174,8 @@ class FeatureSelection(TestCase):
 
         self.assertLess(p_value, self.maximal_p_value_for_significant_features)
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, False,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, False,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertLess(result_series.p_value, self.maximal_p_value_for_significant_features)
         self.assertEqual(result_series.type, "real")
@@ -185,8 +185,8 @@ class FeatureSelection(TestCase):
         x = pd.Series([1.111]*250, name="TEST")
         y = x + pd.Series(np.random.normal(0, 1, 250))
 
-        result_series = tsfresh.feature_selection.feature_selector._calculate_p_value(x, y, False,
-                                                                                      tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
+        result_series = tsfresh.feature_selection.relevance._calculate_p_value(x, y, False,
+                                                                               tsfresh.defaults.TEST_FOR_BINARY_TARGET_REAL_FEATURE)
         self.assertEqual(result_series.name, "TEST")
         self.assertEqual(result_series.type, "const")
         self.assertEqual(result_series.relevant, False)
